third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/directive-parsing-04.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner) {
  testRunner.dumpAsText();
  testRunner.dumpChildFrames();
}
</script>
</head>
<body>
  <p>
    A warning should be logged to the console, as `script-src: 'none'` shouldn't
    contain a colon. Since the directive is invalid, the script should run.
  </p>
  <iframe src="http://127.0.0.1:8000/security/contentSecurityPolicy/resources/echo-script-src.pl?should_run=yes&q=http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js&csp=script-src%3A%20'none'"></iframe>
</body>
</html>
