third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/directive-parsing-04-expected.txt

CONSOLE ERROR: The Content-Security-Policy directive name 'script-src:' contains one or more invalid characters. Only ASCII alphanumeric characters or dashes '-' are allowed in directive names.
A warning should be logged to the console, as `script-src: 'none'` shouldn't contain a colon. Since the directive is invalid, the script should run.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
